requirements.txt
absl-py==2.3.1
accelerate==0.28.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioice==0.10.1
aiortc==1.12.0
aiosignal==1.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
astunparse==1.6.3
attrs==25.4.0
audioread==3.1.0
av==14.4.0
beautifulsoup4==4.14.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
chumpy==0.70
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
crcmod==1.7
cryptography==46.0.3
cycler==0.12.1
Cython==3.2.3
dashscope==1.25.5
decorator==4.4.2
diffusers==0.30.2
distro==1.9.0
dnspython==2.8.0
dynaconf==3.2.12
editdistance==0.8.1
einops==0.8.1
email-validator==2.3.0
fastapi==0.115.14
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
fastrtc @ file:///opt/jyd01/wangruihua/main/OpenAvatarChat_test/src/third_party/gradio_webrtc_videochat/dist/fastrtc-0.0.28.dev0-py3-none-any.whl
ffmpeg-python==0.2.0
ffmpy==0.6.4
filelock==3.20.0
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2024.12.0
funasr==1.2.9
future==1.0.0
gast==0.7.0
gdown==5.2.0
google-crc32c==1.7.1
google-pasta==0.2.0
gradio==5.9.1
gradio_client==1.5.2
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub>=0.34,<1.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.11
ifaddr==0.2.0
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.1
jaconv==0.4.1
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
jiter==0.11.1
jmespath==0.10.0
joblib==1.5.2
json-tricks==3.17.3
kaldiio==2.18.1
keras==3.13.0
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2
llvmlite==0.45.1
loguru==0.7.3
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.8
mdurl==0.1.2
ml-dtypes==0.3.2
mmcv==2.2.0
mmdet==3.1.0
mmengine==0.10.7
mmpose==1.3.2
model-index==0.1.11
modelscope==1.31.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
munkres==1.1.4
namex==0.1.0
networkx==3.5
numba==0.62.1
numpy==1.23.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
onnxruntime==1.20.1
openai==2.6.1
opencv-python==4.12.0.88
opencv-python-headless==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.11
opt_einsum==3.4.0
optree==0.18.0
ordered-set==4.1.0
orjson==3.11.4
oss2==2.19.1
packaging==24.2
pandas==2.3.3
pillow==11.1.0
platformdirs==4.5.0
pooch==1.8.2
proglog==0.1.12
propcache==0.4.1
protobuf==4.25.8
psutil==7.1.3
pyaml==25.7.0
pycocotools==2.0.11
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.2
pylibsrtp==1.0.0
pynndescent==0.5.13
pyOpenSSL==25.3.0
pyparsing==3.3.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytorch-wpe==0.0.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
ruff==0.14.2
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.2.1
sentry-sdk==2.42.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8.1
soxr==1.0.0
starlette==0.41.3
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.16.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.2.0
terminaltables==3.1.10
tf_keras==2.16.0
threadpoolctl==3.6.0
tokenizers==0.19.1
tomlkit==0.13.3
torch==2.8.0+cu128
torch-complex==0.4.4
torchaudio==2.8.0+cu128
torchvision==0.23.0+cu128
tqdm==4.67.1
transformers==4.44.1
triton==3.4.0
typeguard==4.4.4
typer==0.20.0
typer-slim==0.20.1
typing_extensions==4.15.0
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.5.0
uvicorn==0.34.3
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==14.2
Werkzeug==3.1.4
wrapt==2.0.1
xtcocotools==1.14.3
yapf==0.43.0
yarl==1.22.0
zipp==3.23.0
